Create service classes for handling admin shop operations
