+++ src/main/java/com/pigihi/client/auth/AuthPOSTCommunicator.java
@@ -1,5 +1,4 @@
-package com.pigihi.clients.customer;
-
+package com.pigihi.client.auth;
 
 import java.io.IOException;
 import java.net.http.HttpResponse;
@@ -7,42 +6,41 @@ import java.net.http.HttpResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 
-import com.pigihi.clients.ClientCommunicator;
+import com.pigihi.client.ClientCommunicator;
 import com.pigihi.service.POSTRequestSender;
 
-
-public class CustomerPOSTCommunicator implements ClientCommunicator {
-
+public class AuthPOSTCommunicator implements ClientCommunicator {
+	
 	private String endpoint;
 	private String jsonBody = "";
-	
+
+	public AuthPOSTCommunicator(String endpoint) {
+		this.endpoint = endpoint;
+	}
+
+	public AuthPOSTCommunicator(String endpoint, String queryParam, String value) {
+		this.endpoint = endpoint.concat("?").concat(queryParam).concat("=").concat(value);
+	}
+
+	public AuthPOSTCommunicator(String endpoint, String jsonBody) {
+		this.endpoint = endpoint;
+		this.jsonBody = jsonBody;
+	}
+
+	public AuthPOSTCommunicator(String endpoint, String queryParam, String value, String jsonBody) {
+		this.endpoint = endpoint.concat("?").concat(queryParam).concat("=").concat(value);
+		this.jsonBody = jsonBody;
+	}
+
 	@Autowired
 	private POSTRequestSender postRequestSender;
 
-	@Value("${customerService.uri}")
-	private String customerUri;
-
-	public CustomerPOSTCommunicator(String endpoint) {
-		this.endpoint = endpoint;
-	}
-
-	public CustomerPOSTCommunicator(String endpoint, String queryParam, String value) {
-		this.endpoint = endpoint.concat("?").concat(queryParam).concat("=").concat(value);
-	}
-
-	public CustomerPOSTCommunicator(String endpoint, String jsonBody) {
-		this.endpoint = endpoint;
-		this.jsonBody = jsonBody;
-	}
-
-	public CustomerPOSTCommunicator(String endpoint, String queryParam, String value, String jsonBody) {
-		this.endpoint = endpoint.concat("?").concat(queryParam).concat("=").concat(value);
-		this.jsonBody = jsonBody;
-	}
+	@Value("${authService.uri}")
+	private String authUri;
 
 	@Override
 	public String send() throws IOException, InterruptedException {
-		HttpResponse<String> response = postRequestSender.send(customerUri.concat(this.endpoint), this.jsonBody);
+		HttpResponse<String> response = postRequestSender.send(authUri.concat(this.endpoint), this.jsonBody);
 		return response.body();
 	}
