+++ src/main/java/com/pigihi/client/customer/CustomerDELETECommunicator.java
@@ -1,4 +1,4 @@
-package com.pigihi.clients.shop;
+package com.pigihi.client.customer;
 
 import java.io.IOException;
 import java.net.http.HttpResponse;
@@ -6,41 +6,42 @@ import java.net.http.HttpResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 
-import com.pigihi.clients.ClientCommunicator;
+import com.pigihi.client.ClientCommunicator;
+import com.pigihi.service.DELETERequestSender;
 import com.pigihi.service.PATCHRequestSender;
 
-public class ShopPATCHCommunicator implements ClientCommunicator {
+public class CustomerDELETECommunicator implements ClientCommunicator {
 	
 	private String endpoint;
 	private String jsonBody = "";
 	
 	@Autowired
-	private PATCHRequestSender patchRequestSender;
+	private DELETERequestSender deleteRequestSender;
 	
 	@Value("${customerService.uri}")
 	private String customerUri;
-	
-	public ShopPATCHCommunicator(String endpoint) {
+
+	public CustomerDELETECommunicator(String endpoint) {
 		this.endpoint = endpoint;
 	}
 
-	public ShopPATCHCommunicator(String endpoint, String queryParam, String value) {
+	public CustomerDELETECommunicator(String endpoint, String queryParam, String value) {
 		this.endpoint = endpoint.concat("?").concat(queryParam).concat("=").concat(value);
 	}
 
-	public ShopPATCHCommunicator(String endpoint, String jsonBody) {
+	public CustomerDELETECommunicator(String endpoint, String jsonBody) {
 		this.endpoint = endpoint;
 		this.jsonBody = jsonBody;
 	}
 
-	public ShopPATCHCommunicator(String endpoint, String queryParam, String value, String jsonBody) {
+	public CustomerDELETECommunicator(String endpoint, String queryParam, String value, String jsonBody) {
 		this.endpoint = endpoint.concat("?").concat(queryParam).concat("=").concat(value);
 		this.jsonBody = jsonBody;
 	}
 
 	@Override
 	public String send() throws IOException, InterruptedException {
-		HttpResponse<String> response = patchRequestSender.send(customerUri.concat(this.endpoint), this.jsonBody);
+		HttpResponse<String> response = deleteRequestSender.send(customerUri.concat(this.endpoint));
 		return response.body();
 	}
