+++ src/main/java/com/pigihi/client/customer/CustomerGETCommunicator.java
@@ -1,4 +1,5 @@
-package com.pigihi.clients.shop;
+package com.pigihi.client.customer;
+
 
 import java.io.IOException;
 import java.net.http.HttpResponse;
@@ -6,30 +7,31 @@ import java.net.http.HttpResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 
-import com.pigihi.clients.ClientCommunicator;
+import com.pigihi.client.ClientCommunicator;
 import com.pigihi.service.GETRequestSender;
 
-public class ShopGETCommunicator implements ClientCommunicator {
-	
+
+public class CustomerGETCommunicator implements ClientCommunicator {
+
 	private String endpoint;
 	
 	@Autowired
 	private GETRequestSender getRequestSender;
-	
-	@Value("${shopService.uri}")
-	private String shopUri;
-	
-	public ShopGETCommunicator(String endpoint) {
+
+	@Value("${customerService.uri}")
+	private String customerUri;
+
+	public CustomerGETCommunicator(String endpoint) {
 		this.endpoint = endpoint;
 	}
-	
-	public ShopGETCommunicator(String endpoint, String queryParam, String value) {
+
+	public CustomerGETCommunicator(String endpoint, String queryParam, String value) {
 		this.endpoint = endpoint.concat("?").concat(queryParam).concat("=").concat(value);
 	}
 
 	@Override
 	public String send() throws IOException, InterruptedException {
-		HttpResponse<String> response = getRequestSender.send(shopUri.concat(this.endpoint));
+		HttpResponse<String> response = getRequestSender.send(customerUri.concat(this.endpoint));
 		return response.body();
 	}
