+++ src/main/java/com/pigihi/client/customer/CustomerPATCHCommunicator.java
@@ -1,4 +1,4 @@
-package com.pigihi.clients.customer;
+package com.pigihi.client.customer;
 
 import java.io.IOException;
 import java.net.http.HttpResponse;
@@ -6,43 +6,42 @@ import java.net.http.HttpResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 
-import com.pigihi.clients.ClientCommunicator;
-import com.pigihi.service.DELETERequestSender;
+import com.pigihi.client.ClientCommunicator;
 import com.pigihi.service.PATCHRequestSender;
 
-public class CustomerDELETECommunicator implements ClientCommunicator {
+public class CustomerPATCHCommunicator implements ClientCommunicator {
 	
 	private String endpoint;
 	private String jsonBody = "";
 	
 	@Autowired
-	private DELETERequestSender deleteRequestSender;
+	private PATCHRequestSender patchRequestSender;
 	
 	@Value("${customerService.uri}")
 	private String customerUri;
 
-	public CustomerDELETECommunicator(String endpoint) {
+	public CustomerPATCHCommunicator(String endpoint) {
 		this.endpoint = endpoint;
 	}
 
-	public CustomerDELETECommunicator(String endpoint, String queryParam, String value) {
+	public CustomerPATCHCommunicator(String endpoint, String queryParam, String value) {
 		this.endpoint = endpoint.concat("?").concat(queryParam).concat("=").concat(value);
 	}
 
-	public CustomerDELETECommunicator(String endpoint, String jsonBody) {
+	public CustomerPATCHCommunicator(String endpoint, String jsonBody) {
 		this.endpoint = endpoint;
 		this.jsonBody = jsonBody;
 	}
 
-	public CustomerDELETECommunicator(String endpoint, String queryParam, String value, String jsonBody) {
+	public CustomerPATCHCommunicator(String endpoint, String queryParam, String value, String jsonBody) {
 		this.endpoint = endpoint.concat("?").concat(queryParam).concat("=").concat(value);
 		this.jsonBody = jsonBody;
 	}
 
 	@Override
 	public String send() throws IOException, InterruptedException {
-		HttpResponse<String> response = deleteRequestSender.send(customerUri.concat(this.endpoint));
+		HttpResponse<String> response = patchRequestSender.send(customerUri.concat(this.endpoint), this.jsonBody);
 		return response.body();
 	}
-
+	
 }
