+++ src/main/java/com/pigihi/client/shop/ShopPUTCommunicator.java
@@ -1,4 +1,4 @@
-package com.pigihi.clients.shop;
+package com.pigihi.client.shop;
 
 import java.io.IOException;
 import java.net.http.HttpResponse;
@@ -6,7 +6,7 @@ import java.net.http.HttpResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 
-import com.pigihi.clients.ClientCommunicator;
+import com.pigihi.client.ClientCommunicator;
 import com.pigihi.service.PUTRequestSender;
 
 public class ShopPUTCommunicator implements ClientCommunicator {
