+++ src/main/java/com/pigihi/controller/AdminController.java
@@ -16,7 +16,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.pigihi.clients.ClientCommunicator;
+import com.pigihi.client.ClientCommunicator;
 import com.pigihi.entity.AdminEntity;
 import com.pigihi.model.AdminModel;
 import com.pigihi.service.AdminAddService;
