+++ src/main/java/com/pigihi/controller/AdminCustomerController.java
@@ -19,13 +19,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.pigihi.clients.ClientCommunicator;
-import com.pigihi.clients.auth.AuthPOSTCommunicator;
-import com.pigihi.clients.customer.CustomerDELETECommunicator;
-import com.pigihi.clients.customer.CustomerGETCommunicator;
-import com.pigihi.clients.customer.CustomerPATCHCommunicator;
-import com.pigihi.clients.customer.CustomerPOSTCommunicator;
-import com.pigihi.clients.customer.CustomerPUTCommunicator;
+import com.pigihi.client.ClientCommunicator;
+import com.pigihi.client.auth.AuthPOSTCommunicator;
+import com.pigihi.client.customer.CustomerDELETECommunicator;
+import com.pigihi.client.customer.CustomerGETCommunicator;
+import com.pigihi.client.customer.CustomerPATCHCommunicator;
+import com.pigihi.client.customer.CustomerPUTCommunicator;
 import com.pigihi.library.dataConverter.service.DataConverter;
 import com.pigihi.model.CustomerAddressModel;
 import com.pigihi.model.CustomerModel;
