+++ src/main/java/com/pigihi/controller/AdminShopController.java
@@ -18,13 +18,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.pigihi.clients.ClientCommunicator;
-import com.pigihi.clients.auth.AuthPOSTCommunicator;
-import com.pigihi.clients.customer.CustomerDELETECommunicator;
-import com.pigihi.clients.customer.CustomerPATCHCommunicator;
-import com.pigihi.clients.customer.CustomerPUTCommunicator;
-import com.pigihi.clients.shop.ShopGETCommunicator;
-import com.pigihi.clients.shop.ShopPATCHCommunicator;
+import com.pigihi.client.ClientCommunicator;
+import com.pigihi.client.auth.AuthPOSTCommunicator;
+import com.pigihi.client.customer.CustomerDELETECommunicator;
+import com.pigihi.client.customer.CustomerPATCHCommunicator;
+import com.pigihi.client.customer.CustomerPUTCommunicator;
+import com.pigihi.client.shop.ShopGETCommunicator;
+import com.pigihi.client.shop.ShopPATCHCommunicator;
 import com.pigihi.library.dataConverter.service.DataConverter;
 import com.pigihi.model.CustomerModel;
 import com.pigihi.model.EditCustomerModel;
