+++ src/main/java/com/pigihi/service/UserWriteService.java
@@ -2,7 +2,7 @@ package com.pigihi.service;
 
 import java.io.IOException;
 
-import com.pigihi.clients.ClientCommunicator;
+import com.pigihi.client.ClientCommunicator;
 
 public interface UserWriteService {
 	public Object write(ClientCommunicator clientCommunicator) throws IOException, InterruptedException;
