+++ src/main/java/com/pigihi/service/customer/AdminCustomerProfileImageService.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
-import com.pigihi.clients.ClientCommunicator;
+import com.pigihi.client.ClientCommunicator;
 import com.pigihi.service.PUTRequestSender;
 import com.pigihi.service.UserWriteService;
