+++ src/main/java/com/pigihi/service/customer/AdminCustomerStatusService.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
-import com.pigihi.clients.ClientCommunicator;
+import com.pigihi.client.ClientCommunicator;
 import com.pigihi.library.dataConverter.service.DataConverter;
 import com.pigihi.model.EditCustomerModel;
 import com.pigihi.service.DELETERequestSender;
