+++ src/main/java/com/pigihi/service/shop/AdminShopProfileImageService.java
@@ -2,7 +2,7 @@ package com.pigihi.service.shop;
 
 import java.io.IOException;
 
-import com.pigihi.clients.ClientCommunicator;
+import com.pigihi.client.ClientCommunicator;
 import com.pigihi.service.UserWriteService;
 
 public class AdminShopProfileImageService implements UserWriteService {
