+++ src/main/java/com/pigihi/service/shop/AdminShopStatusService.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 
 import org.springframework.stereotype.Service;
 
-import com.pigihi.clients.ClientCommunicator;
+import com.pigihi.client.ClientCommunicator;
 import com.pigihi.service.UserWriteService;
 
 @Service
